Move llama_context setup + perplexity back to main.cpp
